Reworked add and delete endpoints to handle quantities Updated front end App methods to accomodate back end changes Modified Cart Summary Item module to have a quantity field

+++ server/index.js
@@ -33,7 +33,7 @@ app.get('/api/products', (req, res, next) => {
 
 app.get('/api/products/:id', (req, res, next) => {
   const id = req.params.id;
-  if (id <= 0) throw new ClientError(`Product id ${id} is invalid`, 400);
+  if (!Number(id) || id <= 0) throw new ClientError(`Product id ${id} is invalid`, 400);
   db.query(`
     SELECT *
       FROM "products"
@@ -48,42 +48,61 @@ app.get('/api/products/:id', (req, res, next) => {
 
 app.post('/api/cart', (req, res, next) => {
   const pid = req.body.productId;
+  const qty = req.body.quantity;
   if (!pid) throw new ClientError('Product id required', 400);
+  else if (!qty) throw new ClientError('Quantity required', 400);
   else if (!Number(pid) || pid <= 0) throw new ClientError(`Product id ${pid} is invalid`, 400);
+  else if (!Number(qty) || qty <= 0) throw new ClientError(`Quantity ${pid} is invalid`, 400);
   db.query(`
-    SELECT "price"
+    SELECT *
       FROM "products"
      WHERE "productId" = $1;
   `, [pid])
     .then(result => {
       const row = result.rows[0];
       if (!row) throw new ClientError(`Product does not exist at id: ${pid}`, 404);
-      if (req.session.cartId) {
-        return {
-          cartId: req.session.cartId,
-          price: row.price
-        };
-      }
+      if (req.session.cartId) return { cartId: req.session.cartId };
       return db.query(`
         INSERT INTO "carts" ("cartId", "createdAt")
              VALUES (DEFAULT, DEFAULT)
-          RETURNING "cartId", $1 AS "price";
-      `, [row.price]);
+          RETURNING "cartId";
+      `);
     })
     .then(result => {
       const row = result.rows ? result.rows[0] : result;
       req.session.cartId = row.cartId;
       return db.query(`
-        INSERT INTO "cartItems" ("cartId", "productId", "price")
-             VALUES ($1, $2, $3)
-          RETURNING "cartItemId";
-      `, [row.cartId, pid, row.price]);
+        SELECT "cartItemId",
+               "cartId",
+               "productId"
+          FROM "cartItems"
+         WHERE "cartId" = $1
+      `, [row.cartId]);
     })
     .then(result => {
-      const row = result.rows[0];
+      const row = result.rows;
+      const filtered = row.filter(a => Number(pid) === a.productId);
+      if (filtered.length && filtered[0].cartItemId) {
+        return db.query(`
+             UPDATE "cartItems"
+                SET "quantity" = "quantity" + $2
+              WHERE "cartItemId" = $1
+          RETURNING "cartItemId";
+        `, [filtered[0].cartItemId, qty]);
+      } else {
+        return db.query(`
+        INSERT INTO "cartItems" ("cartId", "productId", "quantity")
+             VALUES ($1, $2, $3)
+          RETURNING "cartItemId";
+      `, [req.session.cartId, pid, qty]);
+      }
+    })
+    .then(result => {
+      const row = result.rows ? result.rows[0] : result;
       return db.query(`
         SELECT "c"."cartItemId",
-               "c"."price",
+               "c"."quantity",
+               "p"."price",
                "p"."productId",
                "p"."image",
                "p"."name",
@@ -105,7 +124,7 @@ app.get('/api/cart', (req, res, next) => {
   else {
     db.query(`
       SELECT "c"."cartItemId",
-             "c"."price",
+             "p"."price",
              "p"."productId",
              "p"."image",
              "p"."name",
@@ -130,15 +149,20 @@ app.delete('/api/cart/', (req, res, next) => {
   else if (pid <= 0) throw new ClientError(`Product id ${pid} is invalid`, 400);
   else if (qty <= 0) throw new ClientError(`Quantity ${pid} is invalid`, 400);
   db.query(`
-  DELETE FROM "cartItems"
-        WHERE "cartItemId" IN (
-            SELECT "cartItemId"
-              FROM "cartItems"
-             WHERE "productId" = $1 AND "cartId" = $2
-             ORDER BY "cartItemId" DESC
-             LIMIT $3
-          );
-  `, [pid, cid, qty])
+       UPDATE "cartItems"
+          SET "quantity" = "quantity" - $1
+        WHERE "cartId" = $2 AND "productId" = $3
+    RETURNING "quantity";
+  `, [qty, cid, pid])
+    .then(result => {
+      const row = result.rows[0];
+      if (row.quantity <= 0) {
+        return db.query(`
+          DELETE FROM "cartItems"
+                WHERE "cartId" = $1 AND "productId" = $2
+        `, [cid, pid]);
+      }
+    })
     .then(() => res.sendStatus(204))
     .catch(err => next(err));
 });
@@ -147,8 +171,7 @@ app.post('/api/orders', (req, res, next) => {
   const cid = req.session.cartId;
   const session = req.session;
   const { name, creditCard, shippingAddress } = req.body;
-  if (!cid) throw new ClientError('Cart id required', 400);
-  else if (!name) throw new ClientError('Name required', 400);
+  if (!name) throw new ClientError('Name required', 400);
   else if (!creditCard) throw new ClientError('Credit card required', 400);
   else if (!shippingAddress) throw new ClientError('Shipping address required', 400);
   db.query(`
